[Backend] Refactored routes to report issues and submit suggestions
